Add methods which allows viewing orders

+++ src/main/java/com/example/sklepElektroniczny/entity/Order.java
@@ -4,6 +4,7 @@ import jakarta.persistence.*;
 import jakarta.validation.constraints.Email;
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
@@ -33,10 +34,9 @@ public class Order {
     @JoinColumn(name = "payment_id")
     private Payment payment;
 
-    private Double totalPrice;
+    private BigDecimal totalPrice;
     private String status;
 
-    // Reference to Address
     @ManyToOne
     @JoinColumn(name = "address_id")
     private Address address;
